distributing handlers into separate folders
