+++ lambda/custom/handlers/Channels/ArchiveChannelIntentHandler.js
@@ -1,30 +1,30 @@
 const { ri } = require('@jargon/alexa-skill-sdk');
-const { replaceWhitespacesFunc, login, getRoomId, addAll } = require('../helperFunctions');
+const { replaceWhitespacesFunc, login, getRoomId, archiveChannel } = require('../../helperFunctions');
 
-const AddAllToChannelIntentHandler = {
+const ArchiveChannelIntentHandler = {
 	canHandle(handlerInput) {
 		return handlerInput.requestEnvelope.request.type === 'IntentRequest' &&
-			handlerInput.requestEnvelope.request.intent.name === 'AddAllToChannelIntent';
+			handlerInput.requestEnvelope.request.intent.name === 'ArchiveChannelIntent';
 	},
 	async handle(handlerInput) {
 		try {
 
+			const channelNameData = handlerInput.requestEnvelope.request.intent.slots.archivechannelname.value;
+			const channelName = replaceWhitespacesFunc(channelNameData);
+
 			const {
 				accessToken,
 			} = handlerInput.requestEnvelope.context.System.user;
-			const channelNameData = handlerInput.requestEnvelope.request.intent.slots.addallchannelname.value;
-			const channelName = replaceWhitespacesFunc(channelNameData);
-
 			const headers = await login(accessToken);
 			const roomid = await getRoomId(channelName, headers);
-			const speechText = await addAll(channelName, roomid, headers);
+			const speechText = await archiveChannel(channelName, roomid, headers);
 			const repromptText = ri('GENERIC_REPROMPT');
 
 			return handlerInput.jrb
 				.speak(speechText)
 				.speak(repromptText)
 				.reprompt(repromptText)
-				.withSimpleCard(ri('ADD_ALL_TO_CHANNEL.CARD_TITLE'), speechText)
+				.withSimpleCard(ri('ARCHIVE_CHANNEL.CARD_TITLE'), speechText)
 				.getResponse();
 		} catch (error) {
 			console.error(error);
@@ -33,5 +33,5 @@ const AddAllToChannelIntentHandler = {
 };
 
 module.exports = {
-	AddAllToChannelIntentHandler,
+	ArchiveChannelIntentHandler,
 };
