+++ lambda/custom/handlers/Channels/CreateGrouplIntentHandler.js
@@ -1,30 +1,29 @@
 const { ri } = require('@jargon/alexa-skill-sdk');
-const { replaceWhitespacesFunc, login, getUnreadCounter, channelUnreadMessages } = require('../helperFunctions');
+const { replaceWhitespacesFunc, login, createGroup } = require('../../helperFunctions');
 
-
-const GetUnreadMessagesIntentHandler = {
+const CreateGrouplIntentHandler = {
 	canHandle(handlerInput) {
 		return handlerInput.requestEnvelope.request.type === 'IntentRequest' &&
-			handlerInput.requestEnvelope.request.intent.name === 'ReadUnreadsIntent';
+			handlerInput.requestEnvelope.request.intent.name === 'CreateGroupIntent';
 	},
 	async handle(handlerInput) {
 		try {
 			const {
 				accessToken,
 			} = handlerInput.requestEnvelope.context.System.user;
-			const channelNameData = handlerInput.requestEnvelope.request.intent.slots.readunreadschannel.value;
+
+			const channelNameData = handlerInput.requestEnvelope.request.intent.slots.groupname.value;
 			const channelName = replaceWhitespacesFunc(channelNameData);
 
 			const headers = await login(accessToken);
-			const unreadCount = await getUnreadCounter(channelName, headers);
-			const speechText = await channelUnreadMessages(channelName, unreadCount, headers);
+			const speechText = await createGroup(channelName, headers);
 			const repromptText = ri('GENERIC_REPROMPT');
 
 			return handlerInput.jrb
 				.speak(speechText)
 				.speak(repromptText)
 				.reprompt(repromptText)
-				.withSimpleCard(ri('GET_UNREAD_MESSAGES_FROM_CHANNEL.CARD_TITLE'), speechText)
+				.withSimpleCard(ri('CREATE_CHANNEL.CARD_TITLE'), speechText)
 				.getResponse();
 		} catch (error) {
 			console.error(error);
@@ -33,5 +32,5 @@ const GetUnreadMessagesIntentHandler = {
 };
 
 module.exports = {
-	GetUnreadMessagesIntentHandler,
+	CreateGrouplIntentHandler,
 };
