+++ lambda/custom/handlers/Channels/MakeModeratorIntentHandler.js
@@ -1,16 +1,16 @@
 const { ri } = require('@jargon/alexa-skill-sdk');
-const { replaceWhitespacesDots, replaceWhitespacesFunc, login, getRoomId, getUserId, addOwner } = require('../helperFunctions');
+const { replaceWhitespacesDots, replaceWhitespacesFunc, login, getUserId, getRoomId, makeModerator } = require('../../helperFunctions');
 
-const AddOwnerIntentHandler = {
+const MakeModeratorIntentHandler = {
 	canHandle(handlerInput) {
 		return handlerInput.requestEnvelope.request.type === 'IntentRequest' &&
-			handlerInput.requestEnvelope.request.intent.name === 'AddOwnerIntent';
+			handlerInput.requestEnvelope.request.intent.name === 'MakeModeratorIntent';
 	},
 	async handle(handlerInput) {
 		try {
 
-			const userNameData = handlerInput.requestEnvelope.request.intent.slots.ownerusername.value;
-			const channelNameData = handlerInput.requestEnvelope.request.intent.slots.ownerchannelname.value;
+			const userNameData = handlerInput.requestEnvelope.request.intent.slots.moderatorusername.value;
+			const channelNameData = handlerInput.requestEnvelope.request.intent.slots.moderatorchannelname.value;
 			const userName = replaceWhitespacesDots(userNameData);
 			const channelName = replaceWhitespacesFunc(channelNameData);
 
@@ -20,14 +20,15 @@ const AddOwnerIntentHandler = {
 			const headers = await login(accessToken);
 			const userid = await getUserId(userName, headers);
 			const roomid = await getRoomId(channelName, headers);
-			const speechText = await addOwner(userName, channelName, userid, roomid, headers);
+			const speechText = await makeModerator(userName, channelName, userid, roomid, headers);
 			const repromptText = ri('GENERIC_REPROMPT');
+
 
 			return handlerInput.jrb
 				.speak(speechText)
 				.speak(repromptText)
 				.reprompt(repromptText)
-				.withSimpleCard(ri('ADD_OWNER.CARD_TITLE'), speechText)
+				.withSimpleCard(ri('MAKE_MODERATOR.CARD_TITLE'), speechText)
 				.getResponse();
 		} catch (error) {
 			console.error(error);
@@ -36,5 +37,5 @@ const AddOwnerIntentHandler = {
 };
 
 module.exports = {
-	AddOwnerIntentHandler,
+	MakeModeratorIntentHandler,
 };
