+++ lambda/custom/handlers/Channels/PostMessageIntentHandlers.js
@@ -1,7 +1,7 @@
 const { ri } = require('@jargon/alexa-skill-sdk');
-const { login, resolveChannelname, postMessage } = require('../helperFunctions');
-const { supportsAPL } = require('../utils');
-const titleMessageTemplate = require('../APL/templates/titleMessageTemplate');
+const { login, resolveChannelname, postMessage } = require('../../helperFunctions');
+const { supportsAPL } = require('../../utils');
+const titleMessageTemplate = require('../../APL/templates/titleMessageTemplate');
 
 const StartedPostMessageIntentHandler = {
 	canHandle(handlerInput) {
